Restructure and quality of life
Now the source code, headers (including main.cpp and pch.hpp) and output are in a folder called proj. This not only is more organized and easier to look at, but also fixes the issue where gmake checks for changes to the build's timestamp for some reason, causing it to recompile everything even if no changes were made.

Readded and fixed toolset downloads, also moved them to a new action and made it delete the toolset before redownloading.

Added new action to check the install status, this prints information about the build, all the paths in the configuration file and the status of the toolset (see the actions section of the README.md file for more details.)

Added an script to compile and run in debug automatically.